+++ .idea/misc.xml
@@ -10,15 +10,18 @@
         <entry key="app/src/main/res/drawable/logo_dev.xml" value="0.3355" />
         <entry key="app/src/main/res/drawable/search.xml" value="0.3355" />
         <entry key="app/src/main/res/drawable/settings.xml" value="0.3355" />
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.3111413043478261" />
-        <entry key="app/src/main/res/layout/app_layout.xml" value="0.387037037037037" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.45260416666666664" />
+        <entry key="app/src/main/res/layout/app_layout.xml" value="0.1" />
         <entry key="app/src/main/res/layout/content_main.xml" value="0.3111413043478261" />
+        <entry key="app/src/main/res/layout/fragment_first.xml" value="0.387037037037037" />
         <entry key="app/src/main/res/layout/fragment_history.xml" value="0.387037037037037" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.387037037037037" />
         <entry key="app/src/main/res/layout/fragment_scan.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_second.xml" value="0.3111413043478261" />
         <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.387037037037037" />
         <entry key="app/src/main/res/layout/fragment_shodan.xml" value="0.387037037037037" />
+        <entry key="app/src/main/res/layout/login_fragment.xml" value="0.45260416666666664" />
+        <entry key="app/src/main/res/layout/main_activity.xml" value="0.387037037037037" />
         <entry key="app/src/main/res/layout/test.xml" value="0.4505208333333333" />
         <entry key="app/src/main/res/menu/bottom_nav.xml" value="0.4505208333333333" />
         <entry key="app/src/main/res/menu/bottom_navigation.xml" value="0.4505208333333333" />
