+++ app/src/main/AndroidManifest.xml
@@ -11,13 +11,24 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SecurityIntelligenceToolkit"
+        android:theme="@style/AppFullScreenTheme"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
+            android:label="@string/title_activity_main"
+            android:theme="@style/AppFullScreenTheme">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name=".SITActivity"
+            android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.SecurityIntelligenceToolkit.NoActionBar">
+            android:theme="@style/AppFullScreenTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
