+++ app/src/main/res/layout/app_layout.xml
@@ -17,7 +17,7 @@
             android:id="@+id/topAppBar"
             style="@style/AppButton"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
+            android:layout_height="?android:attr/actionBarSize"
             android:background="@color/purple_500"
             android:visibility="visible"
             app:itemIconTint="@color/white"
@@ -36,7 +36,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
-        android:name="com.toolkit.sit.fragments.HomeFragment"
+        android:name="com.toolkit.sit.fragments.authenticated.HomeFragment"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toTopOf="@+id/bottom_nav"
@@ -59,5 +59,6 @@
         app:layout_constraintRight_toRightOf="parent"
         app:menu="@menu/bottom_nav"
         tools:visibility="visible" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
