+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.toolkit.sit.fragments.SettingsFragment">
+    tools:context="com.toolkit.sit.fragments.authenticated.SettingsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
