+++ app/src/main/res/layout/fragment_shodan.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.toolkit.sit.fragments.ShodanFragment">
+    tools:context="com.toolkit.sit.fragments.authenticated.ShodanFragment">
 
     <TextView
         android:layout_width="match_parent"
