+++ app/src/main/res/layout/login_fragment.xml
@@ -3,13 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.toolkit.sit.fragments.authenticated.HomeFragment">
+    tools:context="com.toolkit.sit.fragments.login.LoginFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:gravity="center"
-        android:text="@string/home"
+        android:text="@string/login"
         android:textSize="60sp" />
 
 </FrameLayout>
