+++ app/src/main/res/values/themes.xml
@@ -14,6 +14,14 @@
         <item name="android:icon">@color/white</item>
         <!-- Customize your theme here. -->
     </style>
+    <style name="AppFullScreenTheme" parent="Theme.MaterialComponents.NoActionBar">
+        <item name="android:windowNoTitle">true</item>
+        <item name="android:windowActionBar">true</item>
+        <item name="android:windowFullscreen">true</item>
+        <item name="android:windowContentOverlay">@null</item>
+    </style>
+
+
 
     <style name="AppButton" parent="Widget.MaterialComponents.Button">
         <item name="backgroundTint">?attr/colorSecondary</item>
@@ -27,4 +35,8 @@
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
+    <style name="Theme.SecurityIntelligenceToolkit.AppBarOverlay"
+        parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+    <style name="Theme.SecurityIntelligenceToolkit.PopupOverlay"
+        parent="ThemeOverlay.AppCompat.Light" />
 </resources>
